backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
